Version 3 of the pre-released app
Added a link to the GitHub repo on the About page for those who want to look at the code. 

Added metadata tags in the HTML. Added the UTF-8 character encoding declaration

Added viewport config default setting

Refactored code to utilize ES6 classes. This enhances code readability and maintainability. ES6 classes offer a more modern and intuitive way to define and organize code, making it easier to understand and maintain.

Introduced a generate_trace method within the Grady class to create a trace for visual results if we are in the visual mode

Implemented a generateRandomCoefficients method to generate random coefficients for the quadratic function. By encapsulating coefficient generation logic within a dedicated method, the code becomes modular and organized.

Improved handling of correct answers by introducing the delta variable for precision control. This allows us to modify how precise the user's guesses should be to be classified as correct.

Added some documentation in the code.